refactor: Break down large components into smaller, reusable components
🎯 Major Refactoring Summary:
- Reduced component complexity by 48-74% across all major files
- Created 5 custom hooks for better logic separation
- Added 13 new reusable UI components
- Improved maintainability and testability

📦 Components Refactored:
- App.tsx: 171 → 59 lines (66% reduction)
- Editor.tsx: 313 → 90 lines (71% reduction)
- Recorder.tsx: 362 → 95 lines (74% reduction)
- ModernStreamingInterface.tsx: 464 → 240 lines (48% reduction)
- Settings.tsx: 224 → 85 lines (62% reduction)

🔧 New Custom Hooks:
- useTourManager: Tour management and notifications
- useProjectManager: Project operations and keyboard shortcuts
- useRecordingManager: Recording state and device management
- useStreamingManager: Streaming state and metrics
- useSettingsManager: Settings persistence and validation

🎨 New UI Components:
- PageTransition: Reusable page transition animations
- MainContent: Main content area with view routing
- EditorHeader, NewProjectDialog, EditorEmptyState
- RecorderHeader, RecordingCompleteNotification, GoLiveButton
- StreamingHeader, StreamingTabNavigation, HelpModal
- SettingsHeader, SettingsSidebar

✅ Benefits:
- Better separation of concerns
- Improved code reusability
- Easier unit testing
- Enhanced maintainability
- Cleaner component architecture

🐛 Fixes:
- Fixed duplicate frameRateOptions declaration
- Fixed invalid Tailwind CSS classes
- Resolved build errors

📝 Updated changelog for tracking changes
