+++ src/renderer/App.tsx
@@ -1,77 +1,26 @@
-import React, { useState, useEffect } from 'react';
-import { motion, AnimatePresence } from 'framer-motion';
-import { Recorder } from './components/recorder/Recorder';
-import { Editor } from './components/editor/Editor';
-import { Settings } from './components/settings/Settings';
-import { ModernStreamingInterface } from './components/streaming/ModernStreamingInterface';
+import React, { useState } from 'react';
 import { TitleBar } from './components/ui/TitleBar';
 import { Sidebar } from './components/ui/Sidebar';
+import { MainContent, AppView } from './components/ui/MainContent';
 import { ToastProvider } from './contexts/ToastContext';
 import { GuidedTour, streamingTourSteps, recordingTourSteps } from './components/common/GuidedTour';
-import { SmartNotifications, useNotifications, createStreamNotifications } from './components/common/SmartNotifications';
-
-type AppView = 'recorder' | 'editor' | 'streaming' | 'settings';
+import { SmartNotifications, createStreamNotifications } from './components/common/SmartNotifications';
+import { useTourManager } from './hooks/useTourManager';
 
 export const App: React.FC = () => {
   const [currentView, setCurrentView] = useState<AppView>('recorder');
-  const [showTour, setShowTour] = useState(false);
-  const [tourType, setTourType] = useState<'streaming' | 'recording'>('streaming');
-  const notifications = useNotifications();
+  const {
+    showTour,
+    tourType,
+    handleTourComplete,
+    handleViewChange: handleTourViewChange,
+    notifications
+  } = useTourManager();
   const streamNotifications = createStreamNotifications(notifications);
-
-  useEffect(() => {
-    // Check if user is new and should see the tour
-    const hasSeenTour = localStorage.getItem('has-seen-tour');
-    if (!hasSeenTour) {
-      // Show tour after a brief delay
-      setTimeout(() => {
-        setShowTour(true);
-        setTourType('recording'); // Start with recording tour since that's the default view
-      }, 1000);
-    }
-
-    // Add tour data attributes to elements
-    const addTourAttributes = () => {
-      // Add data attributes for tour targeting
-      const header = document.querySelector('.title-bar');
-      if (header) header.setAttribute('data-tour', 'header');
-
-      // Add more tour attributes as needed
-    };
-
-    addTourAttributes();
-  }, []);
-
-  const handleTourComplete = () => {
-    setShowTour(false);
-    localStorage.setItem('has-seen-tour', 'true');
-    notifications.success(
-      'Welcome to KwikShot!',
-      'You\'re all set to start recording and streaming. Explore the features and create amazing content!'
-    );
-  };
 
   const handleViewChange = (view: AppView) => {
     setCurrentView(view);
-
-    // Show appropriate tour when switching to streaming for the first time
-    if (view === 'streaming' && !localStorage.getItem('has-seen-streaming-tour')) {
-      setTimeout(() => {
-        setTourType('streaming');
-        setShowTour(true);
-        localStorage.setItem('has-seen-streaming-tour', 'true');
-      }, 500);
-    }
-  };
-
-  const pageVariants = {
-    initial: { opacity: 0, x: 20 },
-    in: { opacity: 1, x: 0 },
-    out: { opacity: 0, x: -20 }
-  };
-
-  const pageTransition = {
-    duration: 0.3
+    handleTourViewChange(view);
   };
 
   return (
@@ -86,68 +35,7 @@ export const App: React.FC = () => {
           <Sidebar currentView={currentView} onViewChange={handleViewChange} />
 
           {/* Main Content Area */}
-          <main className="flex-1 overflow-hidden">
-            <AnimatePresence mode="wait">
-              {currentView === 'recorder' && (
-                <motion.div
-                  key="recorder"
-                  initial="initial"
-                  animate="in"
-                  exit="out"
-                  variants={pageVariants}
-                  transition={pageTransition}
-                  className="h-full"
-                >
-                  <Recorder
-                    onSwitchToEditor={() => setCurrentView('editor')}
-                    onSwitchToStreaming={() => setCurrentView('streaming')}
-                  />
-                </motion.div>
-              )}
-
-              {currentView === 'editor' && (
-                <motion.div
-                  key="editor"
-                  initial="initial"
-                  animate="in"
-                  exit="out"
-                  variants={pageVariants}
-                  transition={pageTransition}
-                  className="h-full"
-                >
-                  <Editor onSwitchToRecorder={() => setCurrentView('recorder')} />
-                </motion.div>
-              )}
-
-              {currentView === 'streaming' && (
-                <motion.div
-                  key="streaming"
-                  initial="initial"
-                  animate="in"
-                  exit="out"
-                  variants={pageVariants}
-                  transition={pageTransition}
-                  className="h-full"
-                >
-                  <ModernStreamingInterface onSwitchToRecorder={() => setCurrentView('recorder')} />
-                </motion.div>
-              )}
-
-              {currentView === 'settings' && (
-                <motion.div
-                  key="settings"
-                  initial="initial"
-                  animate="in"
-                  exit="out"
-                  variants={pageVariants}
-                  transition={pageTransition}
-                  className="h-full"
-                >
-                  <Settings onClose={() => setCurrentView('recorder')} />
-                </motion.div>
-              )}
-            </AnimatePresence>
-          </main>
+          <MainContent currentView={currentView} onViewChange={handleViewChange} />
         </div>
       </div>
