+++ src/renderer/components/editor/NewProjectDialog.tsx
@@ -1,0 +1,77 @@
+import React from 'react';
+import { motion } from 'framer-motion';
+
+interface NewProjectDialogProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onCreateProject: () => void;
+}
+
+export const NewProjectDialog: React.FC<NewProjectDialogProps> = ({
+  isOpen,
+  onClose,
+  onCreateProject
+}) => {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
+      <motion.div
+        className="bg-gray-800 rounded-xl p-6 w-96 border border-gray-700"
+        initial={{ opacity: 0, scale: 0.9 }}
+        animate={{ opacity: 1, scale: 1 }}
+        exit={{ opacity: 0, scale: 0.9 }}
+      >
+        <h3 className="text-lg font-semibold mb-4">Create New Project</h3>
+
+        <div className="space-y-4">
+          <div>
+            <label className="block text-sm text-gray-400 mb-1">Project Name</label>
+            <input
+              type="text"
+              className="input"
+              placeholder="My Video Project"
+              defaultValue="Untitled Project"
+            />
+          </div>
+
+          <div className="grid grid-cols-2 gap-4">
+            <div>
+              <label className="block text-sm text-gray-400 mb-1">Resolution</label>
+              <select className="input text-sm">
+                <option value="1920x1080">1920x1080 (Full HD)</option>
+                <option value="1280x720">1280x720 (HD)</option>
+                <option value="2560x1440">2560x1440 (2K)</option>
+                <option value="3840x2160">3840x2160 (4K)</option>
+              </select>
+            </div>
+
+            <div>
+              <label className="block text-sm text-gray-400 mb-1">Frame Rate</label>
+              <select className="input text-sm">
+                <option value="24">24 FPS</option>
+                <option value="30">30 FPS</option>
+                <option value="60">60 FPS</option>
+              </select>
+            </div>
+          </div>
+        </div>
+
+        <div className="flex justify-end space-x-3 mt-6">
+          <button
+            onClick={onClose}
+            className="btn-ghost"
+          >
+            Cancel
+          </button>
+          <button
+            onClick={onCreateProject}
+            className="btn-primary"
+          >
+            Create Project
+          </button>
+        </div>
+      </motion.div>
+    </div>
+  );
+};
