+++ src/renderer/components/ui/PageTransition.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+
+interface PageTransitionProps {
+  children: React.ReactNode;
+  pageKey: string;
+  className?: string;
+}
+
+export const PageTransition: React.FC<PageTransitionProps> = ({ 
+  children, 
+  pageKey, 
+  className = "h-full" 
+}) => {
+  const pageVariants = {
+    initial: { opacity: 0, x: 20 },
+    in: { opacity: 1, x: 0 },
+    out: { opacity: 0, x: -20 }
+  };
+
+  const pageTransition = {
+    duration: 0.3
+  };
+
+  return (
+    <AnimatePresence mode="wait">
+      <motion.div
+        key={pageKey}
+        initial="initial"
+        animate="in"
+        exit="out"
+        variants={pageVariants}
+        transition={pageTransition}
+        className={className}
+      >
+        {children}
+      </motion.div>
+    </AnimatePresence>
+  );
+};
